% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_classification.R
\name{attribute_sm}
\alias{attribute_sm}
\title{Including SM antecedent for the selected events}
\usage{
attribute_sm(clean_analysis, sm_obs, lag = 1)
}
\arguments{
\item{clean_analysis}{data.frame obtained trough the 'clean_events_analysis' 
function}

\item{sm_obs}{zoo time series of the soil moisture observations}

\item{lag}{the days prior to the first day of precipitation of each event. 
The soil moisture will be obtained from the respective day (P_start - lag)}
}
\value{
data.frame that includes the antecedent soil moisture for each Q and 
P event
}
\description{
Including SM antecedent for the selected events
}
