% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_classification.R
\name{centroid_events}
\alias{centroid_events}
\title{Identify the date of the center of mass of the event and its total volume}
\usage{
centroid_events(observations, events)
}
\arguments{
\item{observations}{zoo time series with the observations}

\item{events}{vector identifying the number of event for the 'observations' 
time series}
}
\description{
Identify the date of the center of mass of the event and its total volume
}
