% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_classification.R
\name{clean_events_analysis}
\alias{clean_events_analysis}
\title{Perform event attribution analysis for P and Q}
\usage{
clean_events_analysis(events_analysis, endP2Qstart_days = 1)
}
\arguments{
\item{events_analysis}{data.frame resulting from the 'match_P2Q_events'
function}

\item{endP2Qstart_days}{numeric value that denotes the maximum lag between 
the end of a P event and start of its respective Q event. P events with lags 
longer to this value will not be considered.}
}
\description{
Perform event attribution analysis for P and Q
}
