% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_classification.R
\name{identify_events}
\alias{identify_events}
\title{Event identification}
\usage{
identify_events(boolean, threshold = 1)
}
\arguments{
\item{boolean}{boolean time series that identify an event (0 = no event, 
1 = event)}

\item{threshold}{values equal to, or lower than this threshold will not be 
counted as an event}
}
\description{
Event identification
}
