% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_classification.R
\name{match_P2Q_events}
\alias{match_P2Q_events}
\title{Attribute Q events to defined P events}
\usage{
match_P2Q_events(p_centroid_events, q_centroid_events)
}
\arguments{
\item{p_centroid_events}{data.frame resulting from the 'centroid_events' 
function applied to P events}

\item{q_centroid_events}{data.frame resulting from the 'centroid_events' 
function applied to Q events}
}
\value{
This function retrieves a data.frame with the respective lag time,
between P and Q events and the respective information of both events 
(intensity, duration; and starting, ending, and CM times)
}
\description{
Attribute Q events to defined P events
}
